refactor: re-structure the /Invite folder, move InvitesSend to InviteEmail, etc
