+++ src/components/Highlighted.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Highlighted(props) {
+  return <span className="red3" {...props} />;
+}
