+++ src/lib/loadingCharacter.js
@@ -1,0 +1,5 @@
+export default val =>
+  val.matchWith({
+    Nothing: () => '▓',
+    Just: p => p.value,
+  });
