+++ src/views/Invite.js
@@ -1,8 +1,9 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import * as need from 'lib/need';
-import { Grid, H5, Text, Button } from 'indigo-react';
+import { Grid, H5, Text } from 'indigo-react';
 
 import { usePointCursor } from 'store/pointCursor';
+import { useHistory } from 'store/history';
 // import { usePointCache } from 'store/pointCache';
 
 import View from 'components/View';
@@ -11,12 +12,12 @@ import Crumbs from 'components/Crumbs';
 import useInvites from 'lib/useInvites';
 import useSyncPoint from 'lib/useSyncPoint';
 import useCurrentPointName from 'lib/useCurrentPointName';
-
-import { useHistory } from 'store/history';
-import { ForwardButton } from 'components/Buttons';
-import FooterButton from 'components/FooterButton';
 import useRouter from 'lib/useRouter';
-import InvitesSend from './InvitesSend';
+import loadingCharacter from 'lib/loadingCharacter';
+
+import InviteUrl from './Invite/InviteUrl';
+import InviteEmail from './Invite/InviteEmail';
+import Highlighted from 'components/Highlighted';
 
 const kInviteNames = {
   URL: 'URL',
@@ -25,48 +26,8 @@ const kInviteNames = {
 
 const kInviteViews = {
   URL: InviteUrl,
-  EMAIL: InvitesSend,
+  EMAIL: InviteEmail,
 };
-
-const valOrLoading = val =>
-  val.matchWith({
-    Nothing: () => '▓',
-    Just: p => p.value,
-  });
-
-function Highlighted(props) {
-  return <span className="red3" {...props} />;
-}
-
-function InviteUrl({ names, push }) {
-  const pushEmail = useCallback(() => push(names.EMAIL), [push, names]);
-
-  return (
-    <>
-      <Grid className="mt3">
-        <Grid.Divider />
-        <Grid.Item
-          as={Button}
-          detail="Generate a URL to send to one guest"
-          full
-          disabled>
-          Generate Invite
-        </Grid.Item>
-        <Grid.Divider />
-        <Grid.Item
-          as={ForwardButton}
-          onClick={pushEmail}
-          detail="Send invitation to your guest's email"
-          full>
-          Send Invite via Email
-        </Grid.Item>
-      </Grid>
-      <FooterButton detail="i.e. Ethereum Address, Text File" disabled>
-        Advanced
-      </FooterButton>
-    </>
-  );
-}
 
 export default function Invite() {
   const history = useHistory();
@@ -84,9 +45,9 @@ export default function Invite() {
 
   // fetch the invites for the current cursor
   const { availableInvites, sentInvites, acceptedInvites } = useInvites(point);
-  const availableInvitesText = valOrLoading(availableInvites);
-  const sentInvitesText = valOrLoading(sentInvites);
-  const acceptedInvitesText = valOrLoading(acceptedInvites);
+  const availableInvitesText = loadingCharacter(availableInvites);
+  const sentInvitesText = loadingCharacter(sentInvites);
+  const acceptedInvitesText = loadingCharacter(acceptedInvites);
 
   return (
     <View>
